src/main/java/com/valdemare/didemo/controllers/GetterInjectedController.java
package com.valdemare.didemo.controllers;

import com.valdemare.didemo.services.GreetingService;
import com.valdemare.didemo.services.GreetingServiceImpl;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Controller;

@Controller
public class GetterInjectedController {

    private GreetingService greetingService;

    public String sayHello(){
        return greetingService.sayGreeting();
    }
    @Autowired
    public void setGreetingService(GreetingService greetingService) {
        this.greetingService = greetingService;
    }
}
